Update the soap-connector version and samples
* Updated soap-connector version to 0.95.1
* Updated the samples to reflect these changes and rebuilt the balx file for the samples
* Changed the package naming of tests for samples from org.wso2.ei.test.connector.soap to org.wso2.ei.sample.test.connector.soap. This is because in the next PR will be adding the package org.wso2.ei.test.connector.soap for connector specific tests.
* Improved the TestUtils class to handle other tests.
