+++ integration-test/test-samples/src/test/java/org/wso2/ei/sample/test/service/orchestration/ServiceOrchestrationTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.ei.test.connector.soap;
+package org.wso2.ei.sample.test.service.orchestration;
 
 import org.apache.commons.httpclient.HttpClient;
 import org.apache.commons.httpclient.methods.PostMethod;
@@ -33,45 +33,43 @@ import org.wso2.ei.test.utils.TestUtils;
 import java.io.IOException;
 
 /**
- * Integration tests for soap connector.
+ * Service orchestration test.
  */
-public class SoapConnectorTest {
-    private static final Logger log = LoggerFactory.getLogger(SoapConnectorTest.class);
+public class ServiceOrchestrationTest {
+    private static final Logger log = LoggerFactory.getLogger(ServiceOrchestrationTest.class);
 
     private boolean serverStarted;
 
     @BeforeClass
     public void startServer() {
-        serverStarted = TestUtils.startServer("samples/soap-connector/soap-connector-sample.balx");
+        serverStarted = TestUtils.startServer("samples/service-orchestration/orchestration.balx");
     }
 
     @Test
-    public void testSoapConnector() {
+    public void testServiceOrchestration() {
         if (!serverStarted) {
             Assert.fail("Error running the test, server is not started");
         }
 
-        String symbol = "ABC";
-        String requestPayload = "<ser:getSimpleQuote xmlns:ser=\"http://services.samples\">\n"
-                + "<ser:symbol>" + symbol + "</ser:symbol>\n"
-                + "</ser:getSimpleQuote>";
+        String vehicleId = "11111";
+        String requestPayload = "{\"Vehicle\":{\"ID\":\"" + vehicleId + "\"\n" + "   },\n" + "   \"card\":{\n"
+                + "      \"no\":\"1234098618781768\",\n" + "      \"cvv\":\"123\"\n" + "   }\n" + "}";
 
-        String url = "http://localhost:9090/soapSample";
+        String url = "http://localhost:9090/license/renew";
 
         HttpClient client = new HttpClient();
         PostMethod post = new PostMethod(url);
 
         try {
-            RequestEntity requestEntity = new StringRequestEntity(requestPayload, "application/xml", "UTF-8");
+            RequestEntity requestEntity = new StringRequestEntity(requestPayload, "application/json", "UTF-8");
 
             post.setRequestEntity(requestEntity);
 
             int status = client.executeMethod(post);
             Assert.assertEquals(status, 200);
             String responseBody = post.getResponseBodyAsString();
-            String responseSymbol = responseBody.split("<ax21:symbol>")[1].split("</ax21:symbol>")[0].trim();
-            Assert.assertEquals(responseSymbol, symbol);
-
+            String responseVehicleId = responseBody.split("Vehicle ID\":\"")[1].split("\"}")[0].trim();
+            Assert.assertEquals(responseVehicleId, vehicleId);
         } catch (IOException e) {
             log.error("Error while invoking the server ", e);
         }
