+++ integration-test/test-samples/src/test/resources/testng.xml
@@ -25,14 +25,14 @@
 
     <test name="Connector tests" verbose="2">
         <classes>
-            <class name="org.wso2.ei.test.connector.soap.SoapConnectorTest"/>
-            <class name="org.wso2.ei.test.connector.file.FileConnectorTest"/>
+            <class name="org.wso2.ei.sample.test.connector.soap.SoapConnectorSampleTest"/>
+            <class name="org.wso2.ei.sample.test.connector.file.FileConnectorTest"/>
         </classes>
     </test>
 
     <test name="Scenario tests" verbose="2">
         <classes>
-            <class name="org.wso2.ei.test.service.orchestration.ServiceOrchestrationTest"/>
+            <class name="org.wso2.ei.sample.test.service.orchestration.ServiceOrchestrationTest"/>
         </classes>
     </test>
