+++ pom.xml
@@ -457,7 +457,7 @@
         <securevault.version>1.0.0-wso2v2</securevault.version>
         <broker.version>1.0.46</broker.version>
 
-        <ballerina.soap.connector.version>0.95.0</ballerina.soap.connector.version>
+        <ballerina.soap.connector.version>0.95.1</ballerina.soap.connector.version>
         <ballerina.file.connector.version>0.95.2</ballerina.file.connector.version>
         <ballerina.jms.connector.version>0.95.4</ballerina.jms.connector.version>
         <ballerina.reliable.delivery.connector.version>0.95.0</ballerina.reliable.delivery.connector.version>
