+++ samples/order-management-system/services/samples/order/GetPaymentDetails.bal
@@ -12,7 +12,7 @@ service<http> sampleSoapServiceForGettingPaymentInfo {
     }
     resource GetPaymentDetails (http:Request request, http:Response response, string orderIdPathParam) {
         endpoint<soap:SoapClient> soapClient {
-            create soap:SoapClient();
+            create soap:SoapClient("http://localhost:9763", {});
         }
 	   
         xmlns "http://ws.starbucks.com" as m0;
@@ -32,7 +32,7 @@ service<http> sampleSoapServiceForGettingPaymentInfo {
         soap:Response soapResponse;
         soap:SoapError soapError;
         soapResponse, soapError = soapClient
-                                  .sendReceive(soapRequest, "http://localhost:9763/services/StarbucksOutletService/");
+                                  .sendReceive("/services/StarbucksOutletService/", soapRequest);
 
         xml backendPayload = soapResponse.payload;
